server: move relay startup flow out of main
